Refactor sensor_service.js and subscriber.js
This commit refactors the code in sensor_service.js and subscriber.js files.
In sensor_service.js, the code has been modified to include a new function called "publishToWebhook" which publishes data to a webhook. Additionally, the "checkImpact" function has been updated to include a payload object that is sent to the webhook when an impact is detected. The "clearBuffer" function has also been modified to reset the buffer array.

In subscriber.js, the code has been modified to export the "publishToWebhook" function. This function publishes data to a webhook by connecting to an MQTT broker and publishing the data to the "webhook" topic.

These changes improve the functionality and maintainability of the code.